+++ examples/abitest/client/cpp/abitest.cc
@@ -24,8 +24,8 @@
 #include "absl/strings/numbers.h"
 #include "absl/strings/str_cat.h"
 #include "absl/strings/string_view.h"
-#include "examples/abitest/client/grpc_test_server.h"
-#include "examples/abitest/client/grpctest.h"
+#include "examples/abitest/client/cpp/grpc_test_server.h"
+#include "examples/abitest/client/cpp/grpctest.h"
 #include "examples/abitest/proto/abitest.grpc.pb.h"
 #include "examples/abitest/proto/abitest.pb.h"
 #include "glog/logging.h"
