+++ examples/abitest/example.toml
@@ -11,8 +11,8 @@ module_0 = { Cargo = { cargo_manifest = "examples/abitest/module_0/rust/Cargo.to
 module_1 = { Cargo = { cargo_manifest = "examples/abitest/module_1/rust/Cargo.toml" } }
 
 [clients]
-cpp = { Bazel = { bazel_target = "//examples/abitest/client:client" }, additional_args = [
+cpp = { Bazel = { bazel_target = "//examples/abitest/client/cpp:client" }, additional_args = [
   "--test_exclude=Storage",
-  "--cert_chain=../../../../../../../../examples/certs/local/local.pem",
-  "--private_key=../../../../../../../../examples/certs/local/local.key",
+  "--cert_chain=../../../../../../../../../examples/certs/local/local.pem",
+  "--private_key=../../../../../../../../../examples/certs/local/local.key",
 ] }
