+++ runner/src/main.rs
@@ -860,7 +860,7 @@ fn run(
                 },
                 "--".to_string(),
                 bazel_target.to_string(),
-                "--ca_cert=../../../../../../../../examples/certs/local/ca.pem".to_string(),
+                "--ca_cert=../../../../../../../../../examples/certs/local/ca.pem".to_string(),
                 ...executable.additional_args.clone(),
                 ...additional_args,
             ],
